requirements.txt
Flask==3.0.3
PyQt5
PyQtWebEngine
arxiv==2.1.3
requests==2.32.3
PyPDF2==3.0.1
scikit-learn==1.5.2
sentence-transformers==3.0.0
nltk==3.9.1
numpy==1.26.4
google-generativeai==0.8.3